# Refactor parseTreeEntries, speed up tree list