This PR will:
* Close #20315
* Speed up #20231


```
$ time git ls-tree -t -r HEAD | wc -l
81941
git ls-tree -t -r HEAD  0.11s user 0.08s system 75% cpu 0.252 total
wc -l  0.01s user 0.00s system 5% cpu 0.251 total

$ time git ls-tree -t -r -l HEAD | wc -l
81941
git ls-tree -t -r -l HEAD  1.00s user 0.77s system 38% cpu 4.578 total
wc -l  0.03s user 0.00s system 0% cpu 4.579 total
```